fix for situation if "which"doesnt exist

+++ releases/sources/term-utils/init.sh
@@ -3,9 +3,7 @@
 path=$1
 
 # Check that ldconfig is exists
-if ! which ldconfig >/dev/null; then
-        cp $path/sbin/ldconfig /sbin/
-fi
+[ -f /sbin/ldconfig ] || cp $path/sbin/ldconfig /sbin/ && /sbin/ldconfig $path/lib
 /sbin/ldconfig $path/lib
 
 [ -f `which htop` ] && rm -f `which htop`
